+++ PrYv-UX-PoC/NotesApp/PYEvent+Helper.h
@@ -16,7 +16,7 @@
 
 
 - (EventDataType)eventDataType;
-- (long)cellStyle;
+- (int)cellStyle;
 
 - (void)firstAttachmentAsImage:(void (^) (UIImage *image))attachmentAsImage
              errorHandler:(void(^) (NSError *error))failure;
